net: qdisc: add a new op to run filters/actions before enqueue
Add new qdisc op 'pre_enqueue' to allow multiqueue queuing disciplines
to call tc_classify or perform other work before dev_pick_tx().

This helps with the default 'mq' queuing discipline and 'mqprio'
queueing discipline where the TXQ effectively picks the qdisc. If
users are trying to put specific flows on a queue it is helpful
to call tc_classify to set the queue_mapping or skb->priority. Once
traffic can be directed to a specific queue then hardware tx rings
may be tuned to support this traffic type. 'mqprio' already gives
the ability to do this via skb->priority the pre_enqueue provides
more control over pkt steering for both mq and mqprio.

Also this allows traffic classifiers to be run without holding
the qdisc lock and gives one place to attach filters when mq
or mqprio are in use. Today a filter has to be attached to each
txq qdisc to have all traffic hit the filter.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>